+++ src/frontend/apps/drive/src/features/explorer/components/embedded-explorer/EmbeddedExplorerGridActionsCell.tsx
@@ -1,6 +1,6 @@
 import { CellContext } from "@tanstack/react-table";
 import { Item, ItemUploadState } from "@/features/drivers/types";
-import { useState } from "react";
+import { memo, useState } from "react";
 import { Button } from "@gouvfr-lasuite/cunningham-react";
 import { Draggable } from "@/features/explorer/components/Draggable";
 import { useDisableDragGridItem } from "./hooks";
@@ -10,7 +10,7 @@ import { useEmbeddedExplorerGirdContext } from "./EmbeddedExplorerGrid";
 
 export type EmbeddedExplorerGridActionsCellProps = CellContext<Item, unknown>;
 
-export const EmbeddedExplorerGridActionsCell = (
+const EmbeddedExplorerGridActionsCellComponent = (
   params: EmbeddedExplorerGridActionsCellProps,
 ) => {
   const item = params.row.original;
@@ -62,3 +62,7 @@ export const EmbeddedExplorerGridActionsCell = (
     </div>
   );
 };
+
+export const EmbeddedExplorerGridActionsCell = memo(
+  EmbeddedExplorerGridActionsCellComponent,
+);