+++ src/frontend/apps/drive/src/features/explorer/components/modals/ExplorerRenameItemModal.tsx
@@ -13,6 +13,7 @@ import { useRef } from "react";
 import { removeFileExtension } from "../../utils/mimeTypes";
 import { useTreeUtils } from "../../hooks/useTreeUtils";
 import { useGlobalExplorer } from "../GlobalExplorerContext";
+import { useSelectionStore } from "../../stores/selectionStore";
 
 type Inputs = {
   title: string;
@@ -26,10 +27,10 @@ export const ExplorerRenameItemModal = (
   const treeUtils = useTreeUtils();
   const {
     rightPanelOpen,
-    selectedItems,
     rightPanelForcedItem,
     setRightPanelForcedItem,
   } = useGlobalExplorer();
+  const selectionStore = useSelectionStore();
   const { t } = useTranslation();
   const form = useForm<Inputs>({
     defaultValues: {
@@ -51,7 +52,8 @@ export const ExplorerRenameItemModal = (
             title: data.title,
           });
 
-          const selectedItem = rightPanelForcedItem ?? selectedItems[0];
+          const selectedItem =
+            rightPanelForcedItem ?? selectionStore.getSelectedItems()[0];
 
           if (rightPanelOpen && selectedItem?.id === props.item.id) {
             const newRightPanelForcedItem = {